added a new abstract class GreedySolver
