+++ cassiopeia/solver/CassiopeiaSolver.py
@@ -13,10 +13,10 @@ from typing import Dict, Optional
 
 class CassiopeiaSolver(abc.ABC):
     """
-  CassiopeiaSolver is an abstract class that all inference algorithms derive
-  from. At minimum, all CassiopeiaSolver subclasses will store a character
-  matrix and implement a solver procedure.
-  """
+    CassiopeiaSolver is an abstract class that all inference algorithms derive
+    from. At minimum, all CassiopeiaSolver subclasses will store a character
+    matrix and implement a solver procedure.
+    """
 
     def __init__(
         self,
@@ -31,6 +31,6 @@ class CassiopeiaSolver(abc.ABC):
 
     @abc.abstractmethod
     def solve(self):
-        """Solve the inference problem
-    """
+        """Solves the inference problem.
+        """
         pass
